refactor: problem service, test service 로직 분리 (#57)
