+++ backend/src/main/java/com/moonspoon/moonspoon/dto/request/test/TestRequest.java
@@ -6,9 +6,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TestRequest {
+    private Long id;
+    private boolean shared;
     private boolean random;
     private int problemCount;
     private String sortOrder;
-
 
 }
