+++ backend/src/main/java/com/moonspoon/moonspoon/problem/ProblemService.java
@@ -1,5 +1,7 @@
 package com.moonspoon.moonspoon.problem;
 
+import com.moonspoon.moonspoon.sharedWorkbook.SharedWorkbook;
+import com.moonspoon.moonspoon.sharedWorkbook.SharedWorkbookRepository;
 import com.moonspoon.moonspoon.workbook.Workbook;
 import com.moonspoon.moonspoon.dto.request.problem.ProblemCreateRequest;
 import com.moonspoon.moonspoon.dto.request.problem.ProblemUpdateRequest;
@@ -36,7 +38,7 @@ import java.util.stream.Collectors;
 public class ProblemService {
     private final ProblemRepository problemRepository;
     private final WorkbookRepository workbookRepository;
-    private Map<String, List<TestResultRequest>> storedLists = new ConcurrentHashMap<>();
+
 
     @Transactional
     public ProblemCreateResponse create(Long workbookId, ProblemCreateRequest dto){
@@ -113,159 +115,6 @@ public class ProblemService {
         problemRepository.delete(problem);
     }
 
-    //Test logic
-    public List<TestProblemResponse> getTestProblems(Long workbookId , TestRequest dto){
 
-        Workbook workbook = validateUserAndWorkbook(workbookId);
-
-        List<Problem> problems = workbook.getProblems();
-
-        int selectCount = Math.min(dto.getProblemCount(), problems.size());
-
-        if(dto.isRandom() && !dto.getSortOrder().equals("none")){
-            //순서 정렬
-            problems = setOrderProblemList(dto.getSortOrder(), problems).subList(0, selectCount);
-            Collections.shuffle(problems);
-        }else if(dto.isRandom() && dto.getSortOrder().equals("none")){
-            Collections.shuffle(problems);
-            problems = problems.subList(0, selectCount);
-        }else if(!dto.isRandom() && !dto.getSortOrder().equals("none")){
-            //순서 정렬
-            problems = setOrderProblemList(dto.getSortOrder(), problems).subList(0, selectCount);
-        }else{
-            //순서 정렬
-            problems = setOrderProblemList("asc", problems).subList(0, selectCount);
-        }
-
-        return problems.stream()
-                .map(p -> TestProblemResponse.fromEntity(p))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> setOrderProblemList(String order, List<Problem> problems){
-        switch(order){
-            case "asc":
-                return sortByCreateDateAsc(problems);
-            case "desc":
-                return sortByCreateDateDesc(problems);
-            case "correctRateAsc":
-                return sortByCorrectRateAsc(problems);
-            case "correctRateDesc":
-                return sortByCorrectRateDesc(problems);
-            default:
-                return problems;
-        }
-    }
-
-    private List<Problem> sortByCorrectRateAsc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparingDouble(Problem::getCorrectRate))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCorrectRateDesc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparingDouble(Problem::getCorrectRate).reversed())
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCreateDateAsc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparing(Problem::getCreateDate))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCreateDateDesc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparing(Problem::getCreateDate).reversed())
-                .collect(Collectors.toList());
-    }
-
-    public List<TestResultResponse> getTestResultProblem(Long workbookId){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        List<TestResultRequest> dto = storedLists.get(username);
-        List<TestResultResponse> responses =  dto.stream()
-                .map(d -> {
-                    Problem problem = problemRepository.findById(d.getId()).orElseThrow(
-                            () -> new NotFoundException("존재하지 않는 문제입니다.")
-                    );
-                    TestResultResponse res = TestResultResponse.fromEntity(problem);
-                    res.setInput(d.getInput());
-                    res.setResult(compareStrings(d.getInput(), problem.getSolution()));
-                    return res;
-                        })
-                .collect(Collectors.toList());
-
-        return responses;
-
-    }
-
-    private String compareStrings(String input, String sol){
-        String inputData = input.replaceAll("\\s", "").toLowerCase();
-        String solution = sol.replaceAll("\\s", "").toLowerCase();
-        if(inputData.equals(solution)){
-            return "correct";
-        }else{
-            return "incorrect";
-        }
-    }
-
-    @Transactional
-    public TestResultSubmitResponse testResultSubmit(Long workbookId, List<TestResultSubmitRequest> dto){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        //계산 및 검증로직
-        int size = dto.size();
-        int correctCount = 0;
-        for(int i = 0; i < dto.size(); i++){
-            if(dto.get(i).getResult().equals("correct")){
-                correctCount++;
-            }
-        }
-        TestResultSubmitResponse response = TestResultSubmitResponse.builder()
-                .correctCount(correctCount)
-                .incorrectCount(size-correctCount)
-                .score(((double)correctCount/size)*100)
-                .problems(dto)
-                .build();
-
-        // 벌크 연산
-        updateCorrectRate(dto);
-
-        return response;
-    }
-
-    private void updateCorrectRate(List<TestResultSubmitRequest> dto){
-        List<Long> problemIds = dto.stream()
-                .map(TestResultSubmitRequest::getId)
-                .collect(Collectors.toList());
-
-        //전체 관련 problem 로드
-        List<Problem> problems = problemRepository.findAllById(problemIds);
-
-        Map<Long, Problem> problemMap = problems.stream()
-                .collect(Collectors.toMap(Problem::getId, Function.identity()));
-
-        for(TestResultSubmitRequest result : dto){
-            Problem problem = problemMap.get(result.getId());
-            if(problem != null){
-                if(result.getResult().equals("correct")){
-                    problem.addCorrectCount();
-                }else{
-                    problem.addIncorrectCount();
-                }
-            }
-        }
-        problemRepository.saveAll(problems);
-    }
-
-    public void storeInputData(Long workbookId ,List<TestResultRequest> listDto){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        this.storedLists.put(username, listDto);
-    }
 
 }
